[SGH-17] Role based accessibility

See merge request sportgh/application!30

+++ src/main/java/edu/codespring/sportgh/controller/ProductController.java
@@ -5,7 +5,9 @@ import edu.codespring.sportgh.dto.ProductOutDTO;
 import edu.codespring.sportgh.dto.ProductPageOutDTO;
 import edu.codespring.sportgh.mapper.ProductMapper;
 import edu.codespring.sportgh.model.Product;
+import edu.codespring.sportgh.security.SecurityUtil;
 import edu.codespring.sportgh.service.ProductService;
+import edu.codespring.sportgh.service.UserService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -25,28 +27,30 @@ public class ProductController {
 
     private final ProductService productService;
     private final ProductMapper productMapper;
+    private final UserService userService;
+    private final SecurityUtil securityUtil;
 
     @GetMapping
     public ResponseEntity<ProductPageOutDTO> findPageByParams(
-            @RequestParam("orderBy") Optional<String> orderBy,
-            @RequestParam("direction") Optional<String> direction,
-            @RequestParam("pageNumber") Optional<Integer> pageNumber,
-            @RequestParam("Subcategory") Optional<String[]> subcategoryNames,
-            @RequestParam("MinPrice") Optional<Double> minPrice,
-            @RequestParam("MaxPrice") Optional<Double> maxPrice,
-            @RequestParam("TextSearch") Optional<String> textSearch
+        @RequestParam("orderBy") Optional<String> orderBy,
+        @RequestParam("direction") Optional<String> direction,
+        @RequestParam("pageNumber") Optional<Integer> pageNumber,
+        @RequestParam("Subcategory") Optional<String[]> subcategoryNames,
+        @RequestParam("MinPrice") Optional<Double> minPrice,
+        @RequestParam("MaxPrice") Optional<Double> maxPrice,
+        @RequestParam("TextSearch") Optional<String> textSearch
     ) {
         return new ResponseEntity<>(
-                productService.findPageByParams(
-                        orderBy.orElse(null),
-                        direction.orElse(null),
-                        pageNumber.orElse(1),
-                        subcategoryNames.orElse(null),
-                        minPrice.orElse(null),
-                        maxPrice.orElse(null),
-                        textSearch.orElse(null)
-                ),
-                HttpStatus.OK
+            productService.findPageByParams(
+                orderBy.orElse(null),
+                direction.orElse(null),
+                pageNumber.orElse(1),
+                subcategoryNames.orElse(null),
+                minPrice.orElse(null),
+                maxPrice.orElse(null),
+                textSearch.orElse(null)
+            ),
+            HttpStatus.OK
         );
 
     }
@@ -63,9 +67,10 @@ public class ProductController {
     @PostMapping
     public ResponseEntity<ProductOutDTO> create(@RequestBody @Valid ProductInDTO productInDTO) {
         log.info("Creating product with name: {}.", productInDTO.getName());
-        if (productInDTO.getId() != null) {
+        if (productInDTO.getUserId() != null) {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
+        productInDTO.setUserId(userService.findByFirebaseUid(productInDTO.getUserUid()).getId());
         return productService.saveInDTO(productInDTO);
     }
 
@@ -95,6 +100,10 @@ public class ProductController {
         Product product = productService.findById(productId);
         if (product == null) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
+
+        if (!securityUtil.isCurrentlyLoggedIn(product.getUser())) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
         }
 
         productService.delete(product);
